% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_candidates.R
\name{adjust_candidates}
\alias{adjust_candidates}
\title{Adjust Data Table with Candidate Elements for Elimination}
\usage{
adjust_candidates(cand, cand_top_idx)
}
\arguments{
\item{cand}{\code{data.table} with candidates for elimination given by their
position indices.}

\item{cand_top_idx}{\code{numeric} position index of candidate to be removed.}
}
\value{
\code{data.table} candidates after any adjustments to position 
indices of candidates.
}
\description{
Adjusts positions of all candidates for elimination in data.table after 
removing a candidate (due to the fact, that the positions may shift).
}
