% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_idx_to_name.R
\name{convert_idx_to_name}
\alias{convert_idx_to_name}
\title{Convert Numbered Index to Named Index of List Element}
\usage{
convert_idx_to_name(vec, obj)
}
\arguments{
\item{vec}{\code{numeric} numeric index of list element.}

\item{obj}{\code{list}}
}
\value{
\code{character} named index of list element.
}
\description{
Convert Numbered Index to Named Index of List Element
}
\examples{
d <- list(a = list(b = list(c = 3, d = 5), e = c(2,4)))
num_idx <- c(1,1,2)
convert_idx_to_name(num_idx, d)
}
