% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_global_variables.R
\name{fix_undefined_global_vars}
\alias{fix_undefined_global_vars}
\title{Fix til at undgå R CMD check notes for "no visible binding for global variable"}
\usage{
fix_undefined_global_vars()
}
\description{
Dette script gør det muligt at referere til kolonner i data frames ved hjælp
af Non Standard Evaluation (NSE) i databehandlingspakker som data.table og
dplyr, uden at dette medfører R CMD check notes angående "no visible binding
for global variable". Navnene på de variable, der refereres til ved hjælp af
NSE, skal blot angives i en vektor til funktionen globalVariables() nedenfor.
}
\details{
Dette er den anbefalede løsning fra CRAN.
}
