% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{pf_obj_size}
\alias{pf_obj_size}
\title{Convert Size in Bytes to Print Friendly String}
\usage{
pf_obj_size(x, digits = 2)
}
\arguments{
\item{x}{\code{numeric} object size in digits.}

\item{digits}{\code{numeric} number of digits you want.}
}
\value{
\code{character} priend friendly string.
}
\description{
Convert Size in Bytes to Print Friendly String
}
\examples{
pf_obj_size(10)
pf_obj_size(1010)
pf_obj_size(2e06)
}
